añadir nuevo archivo borderEffects lucia
