+++ Proyecto CSS/Javascript/index.js
@@ -3,27 +3,9 @@ import { showTime } from "./showTime.js"
 
 document.addEventListener('DOMContentLoaded', showTime)
 
-function borderEffects() {
+import { borderEffects } from "./borderEffects.js"
 
-    const divsWeb = document.querySelectorAll('div, section')
-    console.log(divsWeb)
-
-    for (const div of divsWeb) {
-
-        div.addEventListener('mouseover', (event) => {
-            if (event.altKey) {
-                div.style.border = '1px solid red';
-            }
-        })
-
-        div.addEventListener('mouseout', (event) => {
-            if (event.altKey) {
-                div.style.border = null;
-            }
-        })
-
-    }
-}
+document.addEventListener('load', borderEffect)
 
 function EvaluarCaracter(evObject) {
     let Caracter = String.fromCharCode(evObject.which);
